requirements.txt
# requirements.txt

# huggingface Datasets
datasets

# Core Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
regex>=2023.0.0

# Configuration and Logging
pyyaml>=6.0.0
wandb>=0.15.0
tqdm>=4.65.0

# System Monitoring
psutil>=5.9.0
gputil>=1.4.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Type Checking
mypy>=1.5.1
types-PyYAML>=6.0.12
types-regex>=2023.0.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0

# Optional: Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Development Tools
jupyter>=1.0.0
notebook>=7.0.0
ipython>=8.0.0

# Performance Monitoring
py-spy>=0.3.14

# Data Validation
pydantic>=2.0.0

# Optional: Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Flash Attention
flash-attn